requirements.txt
django>=5.0.7,<5.1.0
djangorestframework
django-cors-headers
pyyaml
requests
